Fix warnings.

+++ fix-gateway-system-tests/src/test/java/uk/co/real_logic/fix_gateway/system_tests/SlowConsumerTest.java
@@ -100,7 +100,7 @@ public class SlowConsumerTest
             testSystem.poll();
         }
 
-        bytesInBufferAtLeast(sessionInfo, (long)senderMaxBytesInBuffer);
+        bytesInBufferAtLeast(sessionInfo, senderMaxBytesInBuffer);
 
         // stopStepping();
     }
